RavenDB-25552 - move to StudioDocumentHandler and ShardedStudioDocumentHandler
