+++ src/Raven.Server/Documents/Sharding/Handlers/ShardedDocumentHandler.cs
@@ -1,6 +1,5 @@
 ﻿using System.Net.Http;
 using System.Threading.Tasks;
-using Raven.Server.Documents.Handlers.Processors.Documents;
 using Raven.Server.Documents.Sharding.Handlers.Processors.Documents;
 using Raven.Server.Routing;
 using Raven.Server.ServerWide.Context;
@@ -22,15 +21,6 @@ namespace Raven.Server.Documents.Sharding.Handlers
         public async Task GetDocSize()
         {
             using (var processor = new ShardedDocumentHandlerProcessorForGetDocSize(this))
-            {
-                await processor.ExecuteAsync();
-            }
-        }
-
-        [RavenShardedAction("/databases/*/docs/validate-schema", "POST")]
-        public async Task ValidateSchema()
-        {
-            using (var processor = new ShardedDocumentHandlerProcessorForValidateDocument(this))
             {
                 await processor.ExecuteAsync();
             }
