+++ src/Raven.Server/Web/Studio/Sharding/ShardedStudioCollectionsHandler.cs
@@ -1,6 +1,5 @@
 ﻿using System.Threading.Tasks;
 using Raven.Server.Documents.Sharding.Handlers;
-using Raven.Server.Documents.Sharding.Handlers.Processors.Revisions;
 using Raven.Server.Documents.Sharding.Handlers.Processors.Studio;
 using Raven.Server.Routing;
 using Raven.Server.Web.Studio.Sharding.Processors;
