+++ src/Raven.Server/Web/Studio/StudioCollectionsHandler.cs
@@ -1,6 +1,5 @@
 ﻿using System.Threading.Tasks;
 using Raven.Server.Documents;
-using Raven.Server.Documents.Handlers.Processors.Revisions;
 using Raven.Server.Documents.Handlers.Processors.Studio;
 using Raven.Server.Routing;
 using Raven.Server.Web.Studio.Processors;
